+++ api/src/features/jobs/use_cases/delete_job.py
@@ -1,6 +1,4 @@
-from restful import use_case as uc
 from services.job_service import JobService
-from starlette.responses import JSONResponse
 
 
 def delete_job_use_case(job_id: str) -> str:
